Moving exceptions to separate package
