+++ core/src/main/java/com/styra/run/discovery/ApiGatewaySelector.java
@@ -5,8 +5,8 @@ import com.styra.run.ApiError;
 import com.styra.run.ApiResponse;
 import com.styra.run.Json;
 import com.styra.run.Result;
-import com.styra.run.StyraRunException;
-import com.styra.run.StyraRunHttpException;
+import com.styra.run.exceptions.StyraRunException;
+import com.styra.run.exceptions.StyraRunHttpException;
 import com.styra.run.utils.Futures;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
