+++ core/src/main/java/com/styra/run/exceptions/RetryException.java
@@ -1,4 +1,4 @@
-package com.styra.run;
+package com.styra.run.exceptions;
 
 public class RetryException extends Exception {
     public RetryException(Throwable cause) {
